+++ C_verzija/sha3_functions.h
@@ -7,6 +7,6 @@
 #define KECCAKF_ROUNDS 24 // 12 + 2*l; l=3
 
 void keccak_f(char* A[MATRIX_DIM][MATRIX_DIM]);
-int32_t rotate_right(int32_t a, const int32_t rotc);
+uint16_t rotate_right(uint16_t a, const uint16_t rotc);
 
 #endif
